# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "description": "Returns list of categories",
                "operationId": "getCategoriesList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Category",
                "description": "category items parent_id,name",
                "operationId": "Category",
                "requestBody": {
                    "description": "Pass category credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parent_id",
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "string",
                                        "format": "string",
                                        "example": "category_name"
                                    },
                                    "parent_id": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "created_at": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "datetime",
                                        "example": "12:34:45 Jun 24 2020"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Store a newly created resource in storage.",
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "name": {
                        "description": "name category",
                        "type": "string",
                        "format": "string",
                        "readOnly": "false"
                    },
                    "parent_id": {
                        "description": "parent_id ",
                        "type": "number",
                        "format": "number",
                        "readOnly": "false"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}